Added example of OAuth and SAML use cases
Documentation only had 1 example which showed how to create application, which isn't really useful when doing production deployments. 

It's much better to have example of SAML and OAuth setups where you set the replyURL to the application and if doing SAML based you add the SAML metadata which allows authentication immediately after you set the application login permissions. Which can't be set from this command unless you have the role setup. So its better to type that in its own tutorial. 

You could add the -GroupMembershipClaims to the example to allow groups, but as the groups are UUID I don't recommend having access to the groups.