API does not allow to choose MailEnabled boolean
API does not allow to choose MailEnabled boolean. If you test for example:

New-AzureADMSGroup -Description "wojtek new" -DisplayName "wojtek new" -MailEnabled $true -MailNickName "wojteknew" -SecurityEnabled $false

You'll be prompt with the following error message: "The service does not currently support writes of mail-enabled groups. Please ensure that the mail-enablement property is unset and the security-enablement property is set."

Looks like we can only create AAD groups with -MailEnabled $false, and so the -SecurityEnabled parameter by consequence cannot be $false, hence it also needs to be $true. Therefore only the -SecurityEnabled $true -MailEnabled $false combination seems to work.

Please update this doc to comply with the current behavior of the cmdlet.