Update List-Service-principal-application-roles.md
Added additional script to the documentation since the script which is already present in the document is good for small tenants with very less applications and assignments, however if there more than 100 assignments per application this script does not work and we would not be able to export all the details. Hence there was a need for additional script which would export all the assignments for a particular service principal in the tenant.